refactor: App 컴포넌트의 라우팅 렌더링 로직 분리
